REFACTOR: Move generic `IdInput` to top-level `components` directory.
This component won't just be used in the model editor.
