+++ packages/frontend/src/model/object_input.tsx
@@ -3,9 +3,9 @@ import { Dynamic } from "solid-js/web";
 import { P, match } from "ts-pattern";
 
 import type { Ob, ObType } from "catlog-wasm";
+import { IdInput, type IdInputOptions } from "../components";
 import type { Theory } from "../theory";
 import { MorphismIndexContext, ObjectIndexContext, TheoryContext } from "./context";
-import { IdInput, type IdInputOptions } from "./id_input";
 
 /** Input an object that already exists in a model.
